requirements.txt
fastapi[standard]
psycopg2-binary
openai
python-dotenv
langgraph
langchain
langchain-community
langchain-core
langchain-openai
pandas
pydantic
numpy
pgvector